requirements.txt
APScheduler==3.7.0
asgiref==3.3.1
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.7
django-apscheduler==0.6.0
django-filter==2.4.0
django-heroku==0.3.1
django-import-export==2.5.0
django-multiupload==0.5.2
django-six==1.0.4
djangorestframework==3.12.2
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.10
jdcal==1.4.1
Markdown==3.3.4
MarkupPy==1.14
micropython-ffilib==0.1.2
numpy==1.21.1
odfpy==1.4.1
openpyxl==3.0.7
oython==1.0.0
pandas==1.3.1
Pillow==8.1.2
psycopg2==2.8.6
pycodestyle==2.7.0
python-dateutil==2.8.2
python-decouple==3.4
python-jsonrpc-server==0.4.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
tzlocal==2.1
ujson==4.0.2
Unipath==1.1
urllib3==1.26.4
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
